Refactor `len_flattened`
